remove `name_contract.rs`
